Allow from_period / to_period to be timestamps
